[ED-450] Update PR template to ensure important checks are done
The dev team hasn't been great at remembering when they need to update documentation, and everybody has a different way of filling out the PR description which often isn't very useful when trying to understand the intent of changes etc.

Suggesting the following updates to the PR template to make it very clear what should be considered, what checks should be done, and when documentation needs to be filled out.

Hopefully this is done in such a way that it makes it easy for citrine to pull out the relevant bullet points and ignore the boilerplate